chocolate, phonons and mutation changes

+++ src/ea/core/mutation.py
@@ -11,9 +11,9 @@ def rearange_mut(work_dir, connections_dir ,atom, lenghts ):
     # optimizer.asu.cell[0][0] = lenghts[0]
     # optimizer.asu.cell[1][1] = lenghts[1]
     # optimizer.asu.cell[2][2] = lenghts[2]
-    optimizer.asu.set_cell([8.211, 5.4624, 65.38, 90, 87.78, 90])
+    # optimizer.asu.set_cell([8.211, 5.4624, 65.38, 90, 87.78, 90])
     try:
-        new = optimizer.run(steps=150, traj=True)
+        new = optimizer.run(steps=300, traj=True)
         best = optimizer.best_structure
     except Exception as e:
         print(e)
@@ -69,7 +69,7 @@ class mutate_reaxff_small():
 
 
 
-    def mutate(self, n_structures, poscar_name='POSCARS_1', factor=0.85):
+    def mutate(self, n_structures, poscar_name='POSCARS_1', factor=0.85, keep_traj=False):
         traj_dir = os.path.join(self.work_dir, self.traj_name)
         trajectory = Trajectory(traj_dir, 'w')
         for i in range(2,int(2+n_structures)):
@@ -93,7 +93,10 @@ class mutate_reaxff_small():
         seeds.poscar_name = poscar_name
         seeds.create(n_structures)
 
-        os.remove(traj_dir)
+        if keep_traj == True:
+            pass
+        else:
+            os.remove(traj_dir)
         shutil.rmtree(os.path.join(self.work_dir, 'CalcFold'))
 
 class mutate_reaxff_small2():
@@ -110,13 +113,15 @@ class mutate_reaxff_small2():
 
         for i in range(2, int(2 + n_structures)):
             atom = self.da.get_atoms(i)
+            sg = self.da.c.get(i).spacegroup
 
             try:
                 optimizer = fix_mol_gradient.GradientDescentGULP(
                     atom,
                     work_dir=self.work_dir,
                     connections=self.connection_dir,
-                    library='reaxff'
+                    library='reaxff',
+                    spacegroup=sg,
                 )
 
                 optimizer.asu.cell *= 0.85  # try modifying cell
@@ -133,7 +138,8 @@ class mutate_reaxff_small2():
                         atom,
                         work_dir=self.work_dir,
                         connections=self.connection_dir,
-                        library='reaxff'
+                        library='reaxff',
+                        spacegroup=sg,
                     )
                     new = optimizer.run(steps=50, traj=True)
                     best = optimizer.best_structure
